+++ flycheck.el
@@ -3810,6 +3810,49 @@ See URL `http://elixir-lang.org/'."
 (defconst flycheck-emacs-args '("-Q" "--batch")
   "Common arguments to Emacs invocations.")
 
+(defun flycheck-emacs-lisp-eval-with-source (form source checker)
+  "Create Emacs arguments to eval FORM with SOURCE for CHECKER.
+
+Create arguments that you can pass to an Emacs subprocess to
+evaluate FORM.
+
+FORM is an s-expression to evaluate.  SOURCE is a string, symbol
+or form suitable for `flycheck-substitute-argument', which see.
+CHECKER is the syntax checker symbol for which FORM is evaluated.
+
+Apply `flycheck-substitute-argument' to SOURCE, wrap a `let'
+expression around FORM, which binds the result of this to the
+variable `flycheck-source' while evaluating FORM, and turn the
+resulting expression into a string with
+`flycheck-sexp-to-string'.
+
+Use this function to pass the file to check to an Emacs Lisp
+subprocess.  Do *not* attempt to pass the file to check as normal
+argument to Emacs.  Even in batch, Emacs will *visit* all files
+specified on the command line.
+
+This can cause various unintended side-effects, ranging from
+unintended loading of libraries or changes to global variables,
+to making the syntax checker process hang, i.e. when Emacs wants
+to ask the user and tries to read the user's answer from standard
+input.  This occurs for instance, if the file to check as unsafe
+local variables.
+
+Instead, use this function to safely splice the file to check
+into the Emacs Lisp form that does the syntax checking.
+
+Return a list of strings, which contains the arguments that can
+be passed to an Emacs executable in order to evaluate FORM with
+the given SOURCE."
+  ;; We use this convoluted way to pass the file check to avoid visiting files.
+  ;; See https://github.com/flycheck/flycheck/issues/319 for an example of the
+  ;; troubles that arise from visiting the file to check
+  (list
+   "--eval"
+   (flycheck-sexp-to-string
+    `(let ((flycheck-source ,(flycheck-substitute-argument source checker)))
+       ,form))))
+
 (defconst flycheck-emacs-lisp-check-form
   '(progn
      (defvar jka-compr-inhibit)
@@ -3821,10 +3864,12 @@ See URL `http://elixir-lang.org/'."
          temp-file))
 
      (setq byte-compile-dest-file-function 'flycheck-byte-compile-dest-file)
+     ;; Flycheck inhibits compression of temporary files, thus we must not
+     ;; attempt to decompress.
      (let ((jka-compr-inhibit t))
-       ;; Flycheck inhibits compression of temporary files, thus we must not
-       ;; attempt to decompress
-       (mapc 'byte-compile-file command-line-args-left))
+       ;; `flycheck-source' is given to us by
+       ;; `flycheck-emacs-lisp-eval-with-source'
+       (byte-compile-file flycheck-source))
      (mapc 'delete-file flycheck-byte-compiled-files)))
 
 (flycheck-def-option-var flycheck-emacs-lisp-load-path nil emacs-lisp
@@ -3906,9 +3951,8 @@ This variable has no effect, if
                     flycheck-option-emacs-lisp-package-user-dir)
             (option "--funcall" flycheck-emacs-lisp-initialize-packages
                     flycheck-option-emacs-lisp-package-initialize)
-            "--eval"
-            (eval (flycheck-sexp-to-string flycheck-emacs-lisp-check-form))
-            source-inplace)
+            (eval (flycheck-emacs-lisp-eval-with-source
+                   flycheck-emacs-lisp-check-form 'source-inplace 'emacs-lisp)))
   :error-patterns
   ((error line-start (file-name) ":" line ":" column ":Error:"
           (message (zero-or-more not-newline)
@@ -3948,15 +3992,14 @@ This variable has no effect, if
 
      (defvar jka-compr-inhibit)
 
-     (let ((filename (car command-line-args-left))
-           ;; Don't attempt to decompress, because Flycheck never compresses the
+     (let (;; Don't attempt to decompress, because Flycheck never compresses the
            ;; temporary files for syntax checking
            (jka-compr-inhibit t)
            ;; Remember the default directory of the process
            (process-default-directory default-directory))
        (with-temp-buffer
-         (insert-file-contents filename 'visit)
-         (setq buffer-file-name filename)
+         (insert-file-contents flycheck-source 'visit)
+         (setq buffer-file-name flycheck-source)
          ;; And change back to the process default directory to make file-name
          ;; back-substutition work
          (setq default-directory process-default-directory)
@@ -3971,9 +4014,9 @@ This variable has no effect, if
   "An Emacs Lisp style checker using CheckDoc.
 
 The checker runs `checkdoc-current-buffer'."
-  :command ("emacs" (eval flycheck-emacs-args) "--eval"
-            (eval (flycheck-sexp-to-string flycheck-emacs-lisp-checkdoc-form))
-            source)
+  :command ("emacs" (eval flycheck-emacs-args)
+            (eval (flycheck-emacs-lisp-eval-with-source
+                   flycheck-emacs-lisp-checkdoc-form 'source 'emacs-lisp-checkdoc)))
   :error-patterns
   ((warning line-start (file-name) ":" line ": " (message) line-end))
   :modes (emacs-lisp-mode)